feat(cmds): trait based commands
This commit adds the abiltity to implement a series of given traits for a command, where each
command should implement **at least** one of these traits; Parsable, ParsableWithArguments,
Creatable. These changed are put in place to ensure that when implementing a command, the proper
methods are implemented in the proper format.