fix(make): don't run certain tasks on workspaces
Previosuly, some tasks such as `run` or `docker-build` would run within all workspaces, this would
cause errors as not all workspaces have binaries and when binary specific things are performed on
the workspaces such as `cargo run`, the entire task would bail. This change just causes those binary
dependent tasks to be run on all workspaces as a whole.