feat(docker): expose .whirl directory to host as a volume
Previously, Whirl would grab the default configuration data from the the `Default` trait of
`Config`, this means that the configuration of a Docker container could not be modified. As well,
the database would be reset upon the restart of the Docker container... This solution introduces a
shared Volume directory with the label `.whirl-data` which acts as a means of persistant data
storage for Whirl.