feat(global): move whirl-generated files to their own directories
This change is in preperations to the new Docker Compose configuration changes that are planned for
the near future. Currently, the Docker Compose configuration does NOT have persistant storage in the
form of Docker Volumes, that will change once the planned updates are pushed. The reason Docker
Volumes might have not been possible (or at least difficult) is due to the fact that Whirl-generated
files were just thrown about, e.g., `whirl.sqlite3`, `Whirl.toml`, all of these files had no central
location so having a shared folder in the form of Docker Volume between the container and the host
wasn't very possible.j

BREAKING CHANGE: Whirl-generated files are moved to their own directories, if the files are not
moved, Whirl will continue to look for these files and panic due to being unable to find them!