feat(cli): fine-grain control over which sub-servers are started
Instead of being given the option to start all of the sub-servers (`whirl run` or `whirl run all`),
you are now given the option to specify a selection of sub-servers by passing a comma-seperated list
to the `run` sub-command (`whirl run distributor,hub`). Despite the changes, `whirl run` still
starts all of the available sub-servers!